docker/requirements.txt
# Basic
jupyterlab
ipykernel
matplotlib
pandas
numpy
scipy
scikit-learn
findspark
pyspark
psycopg2-binary
sockets
